+++ resources/views/components/page/breadcrumbs.blade.php
@@ -1,11 +1,11 @@
 <div>
     @if($model->adminConfig()->getBreadCrumb())
-        <a href="{{ $model->adminConfig()->getBreadCrumb()->url }}" class="link link-primary">
+        <a href="{{ visitedUrl($model->adminConfig()->getBreadCrumb()->url) }}" class="link link-primary">
             <x-chief-icon-label type="back">{{ $model->adminConfig()->getBreadCrumb()->label }}</x-chief-icon-label>
         </a>
     @else
         @adminCan('index')
-            <a href="@adminRoute('index')" class="link link-primary">
+            <a href="{{ visitedUrl($manager->route('index')) }}" class="link link-primary">
                 <x-chief-icon-label type="back">Terug naar overzicht</x-chief-icon-label>
             </a>
         @endAdminCan