+++ resources/views/manager/fields/window/types/select.blade.php
@@ -9,7 +9,7 @@
                 @foreach($fieldGroup['values'] as $optionValue)
                     @if(in_array($optionValue['id'], $selected))
                         <div class="p-0.5">
-                            <span class="inline-block label label-info">{{ $optionValue['label'] }}</span>
+                            <span class="inline-block label label-sm label-info">{{ $optionValue['label'] }}</span>
                         </div>
                     @endif
                 @endforeach
@@ -18,7 +18,7 @@
             @foreach($field->getOptions() as $optionKey => $optionValue)
                 @if(in_array($optionKey, $selected))
                     <div class="p-0.5">
-                        <span class="inline-block label label-info">{{ $optionValue }}</span>
+                        <span class="inline-block label label-sm label-info">{{ $optionValue }}</span>
                     </div>
                 @endif
             @endforeach