+++ CHANGELOG.md
@@ -1,9 +1,26 @@
+
 # Changelog
 
 All Notable changes to the `chief` application template will be documented in this file. Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
-## next release
+# 0.6.15 - 2022-01-27
+- Fixed: pagination on index didn't preserved any filtering or sorting
+- Added: Going back from a detailpage to the index will now preserve any filtering, sorting or paginating on that index page.
+
+## 0.6.14 - 2021-12-24
+- Improved repeatfield to better support custom elements. Now repeat ids are passed to each field so fields can use these unique ids.
+
+## 0.6.12 - 2021-12-14
+- Require php 8.1.
+- Upgraded assetlibrary and medialibrary to version 9. This requires a migration to be run.
+
+## 0.6.9 - 2021-12-08
+- Added: option to set dynamicKeysBlacklist property on a managed model to exclude fields from being dynamic.
+
+## 0.6.8 - 2021-12-06
+- Fixed: squanto window backgrounds layout
+- Added: Field::key() to set custom key
 
 ## 0.6.7 - 2021-11-10
 Removed FieldWindow and FieldSets; customize admin views via blade components